extracted routable-behavior from article-subscriber
